move methods to data manager class
