+++ test/DataManager-test.js
@@ -1,16 +1,14 @@
-import User from '../src/classes/User';
-import Booking from '../src/classes/Booking'
-import Room from '../src/classes/Room'
+import Booking from '../src/classes/Booking';
+import DataManager from '../src/classes/DataManager';
+import Room from '../src/classes/Room';
 
 const chai = require('chai');
 const spies = require('chai-spies');
 const expect = chai.expect;
 chai.use(spies);
 
-describe('User', function() {
-  let user1;
-  let user2;
-  let user3;
+describe.only('DataManager', () => {
+  let dataManager;
   let booking1;
   let booking2;
   let booking3;
@@ -18,18 +16,7 @@ describe('User', function() {
   let room2;
   let room3;
   beforeEach(() => {
-    user1 = new User({
-      'id': 1,
-      'name': 'Leatha Ullrich',
-    });
-    user2 = new User ({
-      "id": 2,
-      "name": "Rocio Schuster"
-    });
-    user3 = new User ({
-      "id": 3,
-      "name": "Kelvin Schiller"
-    })
+    dataManager = new DataManager();
     booking1 = new Booking ({
       "id": "5fwrgu4i7k55hl6sz",
       "userID": 1,
@@ -77,33 +64,22 @@ describe('User', function() {
     })
   })
 
-  it('should be a function', function() {
-    expect(User).to.be.a('function');
+  it('should be able to filter rooms by roomType', () => {
+    const rooms = [room1, room2, room3];
+    const value = "suite"
+
+    const filteredRooms = dataManager.filterRooms(value, rooms)
+
+    expect(filteredRooms).to.deep.equal([room2])
   });
 
-  it('should be able to book a room', function() {
-    user1.bookRoom(room1)
+  it('should be able to find available rooms for a specific date', () => {
+    const rooms = [room1, room2, room3];
+    const bookings = [booking1, booking2, booking3];
+    const date = "2020/01/10";
 
-    expect(user1.roomsBooked.length).to.equal(1)
-    expect(user1.roomsBooked[0]).to.equal(room1)
-  })
+    const availableRooms = dataManager.findAvailableRooms(date, rooms, bookings)
 
-  it('should find the users bookings', function() {
-    global.localStorage = {};
-    chai.spy.on(localStorage, ['setItem', 'getItem'], () => JSON.stringify([booking1, booking2, booking3]));
-
-    const rooms = user1.getBookedHistory()
-
-    expect(rooms).to.deep.equal([booking1, booking3])
-    expect(user1.roomsBooked).to.deep.equal([booking1, booking3])
+    expect(availableRooms).to.deep.equal([room1, room2])
   });
-
-  it('should calculate total spent on rooms', function() {
-    user1.bookRoom(room1)
-    user1.bookRoom(room3)
-
-    expect(user1.calculateRoomCosts()).to.equal(849.54)
-  })
-
-
-})
+});
